# resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "El login exitoso de UserService devuelve un Colaborador con un campo \"token\".\nDicho token se debe incluir como header de las llamadas del resto de las operaciones de cada servicio.",
    "version": "1.0-SNAPSHOT",
    "title": "Sulfur backend",
    "contact": {
      "name": "rsperoni@magnesium.coop"
    }
  },
  "host": "sulfur.mgcoders.com",
  "basePath": "/api",
...
  "schemes": [
    "https"
  ],
  "paths": {
    "/cargos": {
      "get": {
        "tags": [
          "cargos"
        ],
        "summary": "Get cargos",
        "description": "",
        "operationId": "findAll",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cargo"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cargos"
        ],
        "summary": "Create Cargo",
        "description": "",
        "operationId": "create",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Cargo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cargo"
            }
          },
          "500": {
            "description": "Error interno"
          },
          "409": {
            "description": "Código o Id ya existe"
          }
        }
      }
    },
    "/cargos/{id}": {
      "get": {
        "tags": [
          "cargos"
        ],
        "summary": "Get Cargo",
        "description": "",
        "operationId": "find",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cargo"
            }
          },
          "404": {
            "description": "Id no encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "cargos"
        ],
        "summary": "Actualizar precioHora Cargo",
        "description": "",
        "operationId": "actualizarPrecioHora",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/PrecioHora"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cargo"
            }
          },
          "304": {
            "description": "No modificado"
          },
          "409": {
            "description": "Código o Id ya existe"
          }
        }
      },
      "put": {
        "tags": [
          "cargos"
        ],
        "summary": "Edit cargo",
        "description": "",
        "operationId": "edit",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Cargo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Cargo"
            }
          },
          "304": {
            "description": "Error: objeto no modificado"
          }
        }
      }
    },
...
    "/estimaciones": {
      "get": {
        "tags": [
          "estimacion"
        ],
        "summary": "Get estimaciones",
        "description": "",
        "operationId": "findAll",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Estimacion"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "estimacion"
        ],
        "summary": "Create estimacion",
        "description": "",
        "operationId": "create",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Estimacion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Estimacion"
            }
          },
          "400": {
            "description": "Objeto inválido"
          },
          "500": {
            "description": "Error interno"
          },
          "401": {
            "description": "No Autorizado"
          },
          "409": {
            "description": "Código o Id ya existe"
          }
        }
      }
    },
    "/estimaciones/proyecto/{id}": {
      "get": {
        "tags": [
          "estimacion"
        ],
        "summary": "Get estimaciones por proyecto",
        "description": "",
        "operationId": "findAllByProyecto",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Estimacion"
              }
            }
          },
          "400": {
            "description": "Objeto inválido"
          },
          "401": {
            "description": "No Autorizado"
          }
        }
      }
    },
    "/estimaciones/{id}": {
      "get": {
        "tags": [
          "estimacion"
        ],
        "summary": "Get hora",
        "description": "",
        "operationId": "find",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Estimacion"
            }
          },
          "404": {
            "description": "Id no encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "estimacion"
        ],
        "summary": "Edit estimacion",
        "description": "",
        "operationId": "edit",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Estimacion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Estimacion"
            }
          },
          "304": {
            "description": "Error: objeto no modificado"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Cargo": {
      "type": "object",
      "required": [
        "codigo",
        "nombre"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "nombre": {
          "type": "string"
        },
        "codigo": {
          "type": "string"
        },
        "precioHoraHistoria": {
          "type": "array",
          "uniqueItems": true,
          "items": {
            "$ref": "#/definitions/PrecioHora"
          }
        }
      }
    },
    "Proyecto": {
      "type": "object",
      "required": [
        "codigo",
        "nombre"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "codigo": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        }
      }
    },
...
    "EstimacionDetalle": {
      "type": "object",
      "required": [
        "cantidadHoras",
        "cargo",
        "precioTotal",
        "tipoTarea"
      ],
      "properties": {
        "tipoTarea": {
          "$ref": "#/definitions/TipoTarea"
        },
        "cargo": {
          "$ref": "#/definitions/Cargo"
        },
        "cantidadHoras": {
          "type": "integer",
          "format": "int32"
        },
        "precioTotal": {
          "type": "number",
          "format": "double"
        }
      }
    },
...
    "Estimacion": {
      "type": "object",
      "required": [
        "fecha",
        "proyecto"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "proyecto": {
          "$ref": "#/definitions/Proyecto"
        },
        "descripcion": {
          "type": "string"
        },
        "fecha": {
          "type": "string",
          "format": "date",
          "example": "23/01/2017"
        },
        "estimacionDetalleList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EstimacionDetalle"
          }
        }
      }
    },
    "PrecioHora": {
      "type": "object",
      "required": [
        "precioHora",
        "vigenciaDesde"
      ],
      "properties": {
        "precioHora": {
          "type": "number",
          "format": "double"
        },
        "vigenciaDesde": {
          "type": "string",
          "format": "date",
          "example": "23/01/2017"
        }
      }
    },
    "TipoTarea": {
      "type": "object",
      "required": [
        "codigo",
        "nombre"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "codigo": {
          "type": "string"
        },
        "nombre": {
          "type": "string"
        }
      }
    },
...
  }
}